+++ EngineSource/UI/EditorUI/AssetBrowser.h
@@ -17,6 +17,8 @@ public:
 	virtual void OnItemClicked(BrowserItem Item) override;
 	virtual void GoBack() override;
 
+	virtual void OnItemDropped(DroppedItem From, BrowserItem To);
+
 	virtual void DeleteItem(BrowserItem Item) override;
 	static void UpdateAll();
 };