+++ EngineSource/UI/EditorUI/Viewport.cpp
@@ -210,6 +210,8 @@ void Viewport::Tick()
 		}
 	}
 
+	SetName(ChangedScene ? "Viewport*" : "Viewport");
+
 	Vector2 RelativeMouseLocation = Application::GetCursorPosition() - (Position + (Scale * 0.5));
 	Vector3 Rotation = Graphics::MainCamera->ForwardVectorFromScreenPosition(RelativeMouseLocation.X, RelativeMouseLocation.Y);
 
@@ -238,7 +240,7 @@ void Viewport::Tick()
 			for (WorldObject* i : EditorUI::SelectedObjects)
 			{
 				WorldObject* o = Objects::SpawnObjectFromID(i->GetObjectDescription().ID, i->GetTransform());
-				o->SetName(i->GetName());
+				o->Name = i->Name;
 				o->Deserialize(i->Serialize());
 				o->LoadProperties(i->GetPropertiesAsString());
 				o->OnPropertySet();