+++ CHANGES.md
@@ -10,6 +10,7 @@
 - Renamed some C# functions to match their C++ equivalent.
 - Added enum `Input::GamepadType` and `Input::GetGamepadType(Input::Gamepad* From)`.
 - `UIButton` and `UITextField` now inherit from `UIBackground`.
+- Added `Engine.NativeObject` and a few more utility functions to C#.
 
 ## Version 1.7.2